agentctl: move WAL tools to waltools package
This prevents an issue where importing agentctl changed Prometheus
config defaults, breaking converter tests.
