+++ pkg/agentctl/waltools/cardinality_test.go
@@ -1,4 +1,4 @@
-package agentctl
+package waltools
 
 import (
 	"sort"
