+++ pkg/agentctl/waltools/samples.go
@@ -1,4 +1,4 @@
-package agentctl
+package waltools
 
 import (
 	"fmt"
