+++ pkg/agentctl/waltools/walstats_test.go
@@ -1,4 +1,4 @@
-package agentctl
+package waltools
 
 import (
 	"fmt"
